Multi-Modal_Defect_Inspection_and_Visual_QA_System/requirements.txt
streamlit
torch
torchvision
opencv-python
numpy
Pillow
transformers
matplotlib 
albumentations 
numpy 
requests 
tqdm